[REF/#67] ProcedureUiState 내부 data class 분리
